fix typo, move routes to new files
