+++ src/common/constants/routes.json
@@ -1,6 +1,5 @@
 {
-  "LOADING": "/",
-  "DASHBOARD": "/dashboard",
+  "DASHBOARD": "/",
   "SEND": "/send",
   "RECEIVE": "/receive",
   "ADDRESSBOOK": "/addressbook",
