+++ src/common/routes/Routes.tsx
@@ -1,71 +1,29 @@
-import * as React from 'react'
-import { RouteComponentProps } from 'react-router'
-import { Route, Switch, useLocation, useHistory } from 'react-router-dom'
-import { CSSTransition, TransitionGroup } from 'react-transition-group'
+import React from 'react'
+import { Switch, Route } from 'react-router'
+import OnboardingRoutes from './OnboardingRoutes'
+import routes from '../constants/routes.json'
+import Header from '../components/Header'
+import NFTMarketFeed from '../../features/NFTMarket'
+import Dashboard from '../../features/dashboard/DashboardPage'
 
-import { pageRoutes } from './index'
-import * as Styles from './Routes.styles'
-import * as ROUTES from '../utils/constants/routes'
+export default function Routes(): JSX.Element {
+  /**
+   * TODO
+   * Make redux flow for the logged used and avoid passing props
+   */
+  const [user, setUser] = React.useState(true)
 
-type RouteType = {
-  id: string
-  path: string
-  component:
-    | React.FunctionComponent<RouteComponentProps>
-    | React.ComponentClass<RouteComponentProps>
-    | null
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    | any // TODO remove this any and eslint disable after user login is taken by redux
-  layout:
-    | React.FunctionComponent<unknown>
-    | React.ComponentClass<unknown>
-    | null
-}
-
-const childRoutes = (
-  routes: Array<RouteType>,
-  setUser: React.Dispatch<React.SetStateAction<boolean>>,
-) =>
-  routes.map(({ component: Component, layout: Layout, path, id }) => {
-    if (!Component || !Layout) {
-      return null
-    }
-
+  if (user) {
     return (
-      <Route
-        key={id}
-        path={path}
-        exact
-        render={props => (
-          <Layout>
-            <Component {...props} setUser={setUser} />
-          </Layout>
-        )}
-      />
+      <div className='flex flex-col h-full'>
+        <Header />
+        <Switch>
+          <Route exact path={routes.DASHBOARD} component={Dashboard} />
+          <Route exact path={routes.MARKET} component={NFTMarketFeed} />
+        </Switch>
+      </div>
     )
-  })
+  }
 
-interface RoutesProps {
-  setUser: React.Dispatch<React.SetStateAction<boolean>>
+  return <OnboardingRoutes setUser={setUser} />
 }
-
-const Routes: React.FC<RoutesProps> = ({ setUser }) => {
-  const location = useLocation()
-  const history = useHistory()
-
-  React.useEffect(() => history.push(ROUTES.WELCOME_PAGE), [])
-
-  return (
-    <Styles.Container>
-      <TransitionGroup>
-        <CSSTransition key={location.key} classNames='fade' timeout={300}>
-          <Switch location={location}>
-            {childRoutes(pageRoutes, setUser)}
-          </Switch>
-        </CSSTransition>
-      </TransitionGroup>
-    </Styles.Container>
-  )
-}
-
-export default Routes
